requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-appconf==1.0.4
django-countries==7.1
django-crispy-forms==1.11.2
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.12.0
django-relativedelta==1.1.2
django-responsive2==0.1.3
django-sslserver==0.22
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
msgpack==1.0.2
mysqlclient==2.0.3
oauthlib==3.1.0
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
Pillow==8.1.2
pip==21.0.1
psycopg2==2.8.6
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==21.1.0
setuptools==52.0.0.post20210125
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.4
wheel==0.36.2
wincertstore==0.2
zope.interface==5.4.0
